+++ plugins/nf-co2footprint/src/test/nextflow/co2footprint/FileCreation/ReportFileCreatorTest.groovy
@@ -103,10 +103,10 @@ class ReportFileCreatorTest extends Specification{
         totalsJson == totalsJsonResult
         where:
         co2e                || totalsJsonResult
-        10.0d               || [co2e: '10.0 mg', energy:'10.0 mWh', car: '5.71E-5', tree: '28.69s', plane_percent: '2.00E-5 %', plane_flights: null,
-                                co2e_non_cached:'10.0 mg', energy_non_cached:'10.0 mWh', car_non_cached: '5.71E-5', tree_non_cached: '28.69s', plane_percent_non_cached: '2.00E-5 %', plane_flights_non_cached: null]
-        10_000_000_000.0d   || [co2e: '10.0 Mg', energy:'10.0 mWh', car: '5.71E4', tree: '908years 9months 3days 19h 38min 55.87s', plane_percent: null, plane_flights: '200.0',
-                                co2e_non_cached:'10.0 Mg', energy_non_cached:'10.0 mWh', car_non_cached: '5.71E4', tree_non_cached: '908years 9months 3days 19h 38min 55.87s', plane_percent_non_cached: null, plane_flights_non_cached: '200.0']
+        10.0d               || [co2e: '10 mg', energy:'10 mWh', car: '5.71E-5', tree: '28.69s', plane_percent: '2.00E-5 %', plane_flights: null,
+                                co2e_non_cached:'10 mg', energy_non_cached:'10 mWh', car_non_cached: '5.71E-5', tree_non_cached: '28.69s', plane_percent_non_cached: '2.00E-5 %', plane_flights_non_cached: null]
+        10_000_000_000.0d   || [co2e: '10 Mg', energy:'10 mWh', car: '5.71E4', tree: '908years 9months 3days 19h 38min 55.87s', plane_percent: null, plane_flights: '200',
+                                co2e_non_cached:'10 Mg', energy_non_cached:'10 mWh', car_non_cached: '5.71E4', tree_non_cached: '908years 9months 3days 19h 38min 55.87s', plane_percent_non_cached: null, plane_flights_non_cached: '200']
     }
 
     def 'Test payLoad JSON generation' () {
@@ -191,14 +191,14 @@ class ReportFileCreatorTest extends Specification{
         then:
         totalsJson ==
             [
-                co2e: "10.0 mg",
-                energy:  "100.0 mWh",
+                co2e: "10 mg",
+                energy:  "100 mWh",
                 car: "5.71E-5",
                 tree: "28.69s",
                 plane_percent: "2.00E-5 %",
                 plane_flights: null,
-                co2e_non_cached: "10.0 mg",
-                energy_non_cached:  "100.0 mWh",
+                co2e_non_cached: "10 mg",
+                energy_non_cached:  "100 mWh",
                 car_non_cached: "5.71E-5",
                 tree_non_cached: "28.69s",
                 plane_percent_non_cached: '2.00E-5 %',